:car: no bike stations yet
Now we visualise the zone nodes in the network; and generate shortest possible paths between them. I do not use the trips data in here yet.

Note: the graph is currently too large to run the bike placement section on it (unless a better computer can do it or I have adapted it incorrectly).